requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2.7
django-debug-toolbar==3.7.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
phonenumbers==8.12.56
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
XlsxWriter==3.0.3
